Put all invariants in right order
