+++ src/controller_examples/rabbitmq_controller/proof/liveness/liveness.rs
@@ -1,6 +1,7 @@
 // Copyright 2022 VMware, Inc.
 // SPDX-License-Identifier: MIT
 #![allow(unused_imports)]
+use super::{helper_invariants, terminate};
 use crate::external_api::spec::*;
 use crate::kubernetes_api_objects::{
     api_method::*, common::*, config_map::*, dynamic::*, owner_reference::*, resource::*,
@@ -15,7 +16,7 @@ use crate::kubernetes_cluster::spec::{
 };
 use crate::rabbitmq_controller::{
     common::*,
-    proof::{common::*, liveness::helper_invariants, terminate},
+    proof::common::*,
     spec::{rabbitmqcluster::*, reconciler::*},
 };
 use crate::temporal_logic::{defs::*, rules::*};
@@ -149,7 +150,7 @@ spec fn invariants_phase_I(rabbitmq: RabbitmqClusterView) -> TempPred<RMQCluster
     .and(always(lift_state(RMQCluster::each_scheduled_key_is_consistent_with_its_object())))
     .and(always(lift_state(RMQCluster::each_key_in_reconcile_is_consistent_with_its_object())))
     .and(always(lift_state(helper_invariants::stateful_set_has_no_finalizers_or_timestamp_and_only_has_controller_owner_ref(rabbitmq))))
-    .and(always(lift_state(helper_invariants::server_config_map_has_no_finalizers_or_timestamp(rabbitmq))))
+    .and(always(lift_state(helper_invariants::server_config_map_has_no_finalizers_or_timestamp_and_only_has_controller_owner_ref(rabbitmq))))
     .and(always(lift_state(helper_invariants::pending_msg_at_after_create_server_config_map_step_is_create_cm_req(rabbitmq.object_ref()))))
     .and(always(lift_state(helper_invariants::pending_msg_at_after_update_server_config_map_step_is_update_cm_req(rabbitmq.object_ref()))))
     .and(always(lift_state(helper_invariants::pending_msg_at_after_create_stateful_set_step_is_create_sts_req(rabbitmq.object_ref()))))
@@ -184,7 +185,7 @@ proof fn invariants_is_stable(rabbitmq: RabbitmqClusterView)
         lift_state(RMQCluster::each_scheduled_key_is_consistent_with_its_object()),
         lift_state(RMQCluster::each_key_in_reconcile_is_consistent_with_its_object()),
         lift_state(helper_invariants::stateful_set_has_no_finalizers_or_timestamp_and_only_has_controller_owner_ref(rabbitmq)),
-        lift_state(helper_invariants::server_config_map_has_no_finalizers_or_timestamp(rabbitmq)),
+        lift_state(helper_invariants::server_config_map_has_no_finalizers_or_timestamp_and_only_has_controller_owner_ref(rabbitmq)),
         lift_state(helper_invariants::pending_msg_at_after_create_server_config_map_step_is_create_cm_req(rabbitmq.object_ref())),
         lift_state(helper_invariants::pending_msg_at_after_update_server_config_map_step_is_update_cm_req(rabbitmq.object_ref())),
         lift_state(helper_invariants::pending_msg_at_after_create_stateful_set_step_is_create_sts_req(rabbitmq.object_ref())),
@@ -245,8 +246,8 @@ proof fn invariants_phase_III_is_stable(rabbitmq: RabbitmqClusterView)
 spec fn invariants_phase_IV(rabbitmq: RabbitmqClusterView) -> TempPred<RMQCluster> {
     always(lift_state(helper_invariants::at_most_one_create_cm_req_is_in_flight(rabbitmq.object_ref())))
     .and(always(lift_state(helper_invariants::at_most_one_update_cm_req_is_in_flight(rabbitmq.object_ref()))))
-    .and(always(lift_state(helper_invariants::no_delete_cm_req_is_in_flight(rabbitmq.object_ref()))))
     .and(always(lift_state(helper_invariants::every_update_cm_req_does_the_same(rabbitmq))))
+    .and(always(lift_state(helper_invariants::every_create_cm_req_does_the_same(rabbitmq))))
     .and(always(lift_state(helper_invariants::at_most_one_create_sts_req_is_in_flight(rabbitmq.object_ref()))))
     .and(always(lift_state(helper_invariants::at_most_one_update_sts_req_is_in_flight(rabbitmq.object_ref()))))
     .and(always(lift_state(helper_invariants::every_update_sts_req_does_the_same(rabbitmq))))
@@ -260,8 +261,8 @@ proof fn invariants_phase_IV_is_stable(rabbitmq: RabbitmqClusterView)
     stable_and_always_n!(
         lift_state(helper_invariants::at_most_one_create_cm_req_is_in_flight(rabbitmq.object_ref())),
         lift_state(helper_invariants::at_most_one_update_cm_req_is_in_flight(rabbitmq.object_ref())),
-        lift_state(helper_invariants::no_delete_cm_req_is_in_flight(rabbitmq.object_ref())),
         lift_state(helper_invariants::every_update_cm_req_does_the_same(rabbitmq)),
+        lift_state(helper_invariants::every_create_cm_req_does_the_same(rabbitmq)),
         lift_state(helper_invariants::at_most_one_create_sts_req_is_in_flight(rabbitmq.object_ref())),
         lift_state(helper_invariants::at_most_one_update_sts_req_is_in_flight(rabbitmq.object_ref())),
         lift_state(helper_invariants::every_update_sts_req_does_the_same(rabbitmq)),
@@ -273,24 +274,32 @@ proof fn invariants_phase_IV_is_stable(rabbitmq: RabbitmqClusterView)
 // Different from above, it only holds after some time since the rest id counter is the same as rest_id.
 spec fn invariants_phase_V(rabbitmq: RabbitmqClusterView) -> TempPred<RMQCluster> {
     always(lift_state(helper_invariants::stateful_set_has_owner_reference_pointing_to_current_cr(rabbitmq)))
+    .and(always(lift_state(helper_invariants::server_config_map_has_owner_reference_pointing_to_current_cr(rabbitmq))))
 }
 
 proof fn invariants_phase_V_is_stable(rabbitmq: RabbitmqClusterView)
     ensures
         valid(stable(invariants_phase_V(rabbitmq))),
 {
-    always_p_is_stable(lift_state(helper_invariants::stateful_set_has_owner_reference_pointing_to_current_cr(rabbitmq)));
+    stable_and_always_n!(
+        lift_state(helper_invariants::stateful_set_has_owner_reference_pointing_to_current_cr(rabbitmq)),
+        lift_state(helper_invariants::server_config_map_has_owner_reference_pointing_to_current_cr(rabbitmq))
+    );
 }
 
 spec fn invariants_phase_VI(rabbitmq: RabbitmqClusterView) -> TempPred<RMQCluster> {
     always(lift_state(helper_invariants::no_delete_sts_req_is_in_flight(rabbitmq.object_ref())))
+    .and(always(lift_state(helper_invariants::no_delete_cm_req_is_in_flight(rabbitmq.object_ref()))))
 }
 
 proof fn invariants_phase_VI_is_stable(rabbitmq: RabbitmqClusterView)
     ensures
         valid(stable(invariants_phase_VI(rabbitmq))),
 {
-    always_p_is_stable(lift_state(helper_invariants::no_delete_sts_req_is_in_flight(rabbitmq.object_ref())));
+    stable_and_always_n!(
+        lift_state(helper_invariants::no_delete_sts_req_is_in_flight(rabbitmq.object_ref())),
+        lift_state(helper_invariants::no_delete_cm_req_is_in_flight(rabbitmq.object_ref()))
+    );
 }
 
 proof fn liveness_proof(rabbitmq: RabbitmqClusterView)
@@ -327,6 +336,16 @@ proof fn liveness_proof(rabbitmq: RabbitmqClusterView)
             temp_pred_equality(true_pred().and(invariants_phase_VI(rabbitmq)), invariants_phase_VI(rabbitmq));
             eliminate_always(spec, lift_state(RMQCluster::every_in_flight_msg_has_lower_id_than_allocator()));
             helper_invariants::lemma_true_leads_to_always_no_delete_sts_req_is_in_flight(spec, rabbitmq.object_ref());
+            helper_invariants::lemma_true_leads_to_always_no_delete_cm_req_is_in_flight(spec, rabbitmq.object_ref());
+            leads_to_always_combine_temp(
+                spec, true_pred(),
+                lift_state(helper_invariants::no_delete_sts_req_is_in_flight(rabbitmq.object_ref())),
+                lift_state(helper_invariants::no_delete_cm_req_is_in_flight(rabbitmq.object_ref()))
+            );
+            always_and_equality(
+                lift_state(helper_invariants::no_delete_sts_req_is_in_flight(rabbitmq.object_ref())),
+                lift_state(helper_invariants::no_delete_cm_req_is_in_flight(rabbitmq.object_ref()))
+            );
             leads_to_trans_temp(spec, true_pred(), invariants_phase_VI(rabbitmq), always(current_state_matches(rabbitmq)));
         }
     );
@@ -344,8 +363,18 @@ proof fn liveness_proof(rabbitmq: RabbitmqClusterView)
                 invariants_phase_IV(rabbitmq)
             );
             unpack_conditions_from_spec(spec, invariants_phase_V(rabbitmq), true_pred(), always(current_state_matches(rabbitmq)));
-            temp_pred_equality(true_pred().and(invariants_phase_V(rabbitmq)), always(lift_state(helper_invariants::stateful_set_has_owner_reference_pointing_to_current_cr(rabbitmq))));
-            lemma_eventually_only_valid_stateful_set_exists(spec, rabbitmq);
+            temp_pred_equality(true_pred().and(invariants_phase_V(rabbitmq)), invariants_phase_V(rabbitmq));
+            helper_invariants::lemma_eventually_only_valid_stateful_set_exists(spec, rabbitmq);
+            helper_invariants::lemma_eventually_only_valid_server_config_map_exists(spec, rabbitmq);
+            leads_to_always_combine_temp(
+                spec, true_pred(),
+                lift_state(helper_invariants::stateful_set_has_owner_reference_pointing_to_current_cr(rabbitmq)),
+                lift_state(helper_invariants::server_config_map_has_owner_reference_pointing_to_current_cr(rabbitmq))
+            );
+            always_and_equality(
+                lift_state(helper_invariants::stateful_set_has_owner_reference_pointing_to_current_cr(rabbitmq)),
+                lift_state(helper_invariants::server_config_map_has_owner_reference_pointing_to_current_cr(rabbitmq))
+            );
             leads_to_trans_temp(spec, true_pred(), invariants_phase_V(rabbitmq), always(current_state_matches(rabbitmq)));
         }
     );
@@ -372,8 +401,8 @@ proof fn liveness_proof(rabbitmq: RabbitmqClusterView)
 
             helper_invariants::lemma_true_leads_to_always_at_most_one_create_cm_req_is_in_flight(spec, rabbitmq.object_ref());
             helper_invariants::lemma_true_leads_to_always_at_most_one_update_cm_req_is_in_flight(spec, rabbitmq.object_ref());
-            helper_invariants::lemma_true_leads_to_always_no_delete_cm_req_is_in_flight(spec, rabbitmq.object_ref());
             helper_invariants::lemma_true_leads_to_always_every_update_cm_req_does_the_same(spec, rabbitmq);
+            helper_invariants::lemma_true_leads_to_always_every_create_cm_req_does_the_same(spec, rabbitmq);
             helper_invariants::lemma_true_leads_to_always_at_most_one_create_sts_req_is_in_flight(spec, rabbitmq.object_ref());
             helper_invariants::lemma_true_leads_to_always_at_most_one_update_sts_req_is_in_flight(spec, rabbitmq.object_ref());
             helper_invariants::lemma_true_leads_to_always_every_update_sts_req_does_the_same(spec, rabbitmq);
@@ -383,8 +412,8 @@ proof fn liveness_proof(rabbitmq: RabbitmqClusterView)
                 spec, true_pred(),
                 lift_state(helper_invariants::at_most_one_create_cm_req_is_in_flight(rabbitmq.object_ref())),
                 lift_state(helper_invariants::at_most_one_update_cm_req_is_in_flight(rabbitmq.object_ref())),
-                lift_state(helper_invariants::no_delete_cm_req_is_in_flight(rabbitmq.object_ref())),
                 lift_state(helper_invariants::every_update_cm_req_does_the_same(rabbitmq)),
+                lift_state(helper_invariants::every_create_cm_req_does_the_same(rabbitmq)),
                 lift_state(helper_invariants::at_most_one_create_sts_req_is_in_flight(rabbitmq.object_ref())),
                 lift_state(helper_invariants::at_most_one_update_sts_req_is_in_flight(rabbitmq.object_ref())),
                 lift_state(helper_invariants::every_update_sts_req_does_the_same(rabbitmq)),
@@ -394,8 +423,8 @@ proof fn liveness_proof(rabbitmq: RabbitmqClusterView)
             always_and_equality_n!(
                 lift_state(helper_invariants::at_most_one_create_cm_req_is_in_flight(rabbitmq.object_ref())),
                 lift_state(helper_invariants::at_most_one_update_cm_req_is_in_flight(rabbitmq.object_ref())),
-                lift_state(helper_invariants::no_delete_cm_req_is_in_flight(rabbitmq.object_ref())),
                 lift_state(helper_invariants::every_update_cm_req_does_the_same(rabbitmq)),
+                lift_state(helper_invariants::every_create_cm_req_does_the_same(rabbitmq)),
                 lift_state(helper_invariants::at_most_one_create_sts_req_is_in_flight(rabbitmq.object_ref())),
                 lift_state(helper_invariants::at_most_one_update_sts_req_is_in_flight(rabbitmq.object_ref())),
                 lift_state(helper_invariants::every_update_sts_req_does_the_same(rabbitmq)),
@@ -506,7 +535,7 @@ proof fn sm_spec_entails_all_invariants(rabbitmq: RabbitmqClusterView)
     RMQCluster::lemma_always_each_scheduled_key_is_consistent_with_its_object(spec);
     RMQCluster::lemma_always_each_key_in_reconcile_is_consistent_with_its_object(spec);
     helper_invariants::lemma_always_stateful_set_has_no_finalizers_or_timestamp_and_only_has_controller_owner_ref(spec, rabbitmq);
-    helper_invariants::lemma_always_server_config_map_has_no_finalizers_or_timestamp(spec, rabbitmq);
+    helper_invariants::lemma_always_server_config_map_has_no_finalizers_or_timestamp_and_only_has_controller_owner_ref(spec, rabbitmq);
     helper_invariants::lemma_always_pending_msg_at_after_create_server_config_map_step_is_create_cm_req(spec, rabbitmq.object_ref());
     helper_invariants::lemma_always_pending_msg_at_after_update_server_config_map_step_is_update_cm_req(spec, rabbitmq.object_ref());
     helper_invariants::lemma_always_pending_msg_at_after_create_stateful_set_step_is_create_sts_req(spec, rabbitmq.object_ref());
@@ -564,7 +593,7 @@ proof fn sm_spec_entails_all_invariants(rabbitmq: RabbitmqClusterView)
         lift_state(RMQCluster::each_scheduled_key_is_consistent_with_its_object()),
         lift_state(RMQCluster::each_key_in_reconcile_is_consistent_with_its_object()),
         lift_state(helper_invariants::stateful_set_has_no_finalizers_or_timestamp_and_only_has_controller_owner_ref(rabbitmq)),
-        lift_state(helper_invariants::server_config_map_has_no_finalizers_or_timestamp(rabbitmq)),
+        lift_state(helper_invariants::server_config_map_has_no_finalizers_or_timestamp_and_only_has_controller_owner_ref(rabbitmq)),
         lift_state(helper_invariants::pending_msg_at_after_create_server_config_map_step_is_create_cm_req(rabbitmq.object_ref())),
         lift_state(helper_invariants::pending_msg_at_after_update_server_config_map_step_is_update_cm_req(rabbitmq.object_ref())),
         lift_state(helper_invariants::pending_msg_at_after_create_stateful_set_step_is_create_sts_req(rabbitmq.object_ref())),
@@ -1384,7 +1413,7 @@ proof fn lemma_from_after_get_stateful_set_and_key_exists_to_rabbitmq_matches(ra
                     let object = ex.head().resource_obj_of(make_stateful_set_key(rabbitmq.object_ref()));
                     assert(pre_with_object(object).satisfied_by(ex));
                 }
-                
+
             }
         }
     );
@@ -2577,7 +2606,7 @@ proof fn lemma_cm_is_updated_at_after_update_server_config_map_step_with_rabbitm
         spec.entails(always(lift_state(RMQCluster::each_object_in_etcd_is_well_formed()))),
         spec.entails(always(lift_state(helper_invariants::at_most_one_update_cm_req_is_in_flight(rabbitmq.object_ref())))),
         spec.entails(always(lift_state(helper_invariants::no_delete_cm_req_is_in_flight(rabbitmq.object_ref())))),
-        spec.entails(always(lift_state(helper_invariants::server_config_map_has_no_finalizers_or_timestamp(rabbitmq)))),
+        spec.entails(always(lift_state(helper_invariants::server_config_map_has_no_finalizers_or_timestamp_and_only_has_controller_owner_ref(rabbitmq)))),
         rabbitmq.well_formed(),
     ensures
         spec.entails(
@@ -2616,7 +2645,7 @@ proof fn lemma_cm_is_updated_at_after_update_server_config_map_step_with_rabbitm
         &&& RMQCluster::each_object_in_etcd_is_well_formed()(s)
         &&& helper_invariants::at_most_one_update_cm_req_is_in_flight(rabbitmq.object_ref())(s)
         &&& helper_invariants::no_delete_cm_req_is_in_flight(rabbitmq.object_ref())(s)
-        &&& helper_invariants::server_config_map_has_no_finalizers_or_timestamp(rabbitmq)(s)
+        &&& helper_invariants::server_config_map_has_no_finalizers_or_timestamp_and_only_has_controller_owner_ref(rabbitmq)(s)
     };
     entails_always_and_n!(
         spec,
@@ -2627,7 +2656,7 @@ proof fn lemma_cm_is_updated_at_after_update_server_config_map_step_with_rabbitm
         lift_state(RMQCluster::each_object_in_etcd_is_well_formed()),
         lift_state(helper_invariants::at_most_one_update_cm_req_is_in_flight(rabbitmq.object_ref())),
         lift_state(helper_invariants::no_delete_cm_req_is_in_flight(rabbitmq.object_ref())),
-        lift_state(helper_invariants::server_config_map_has_no_finalizers_or_timestamp(rabbitmq))
+        lift_state(helper_invariants::server_config_map_has_no_finalizers_or_timestamp_and_only_has_controller_owner_ref(rabbitmq))
     );
     temp_pred_equality(
         lift_action(stronger_next),
@@ -2638,7 +2667,7 @@ proof fn lemma_cm_is_updated_at_after_update_server_config_map_step_with_rabbitm
         .and(lift_state(RMQCluster::each_object_in_etcd_is_well_formed()))
         .and(lift_state(helper_invariants::at_most_one_update_cm_req_is_in_flight(rabbitmq.object_ref())))
         .and(lift_state(helper_invariants::no_delete_cm_req_is_in_flight(rabbitmq.object_ref())))
-        .and(lift_state(helper_invariants::server_config_map_has_no_finalizers_or_timestamp(rabbitmq)))
+        .and(lift_state(helper_invariants::server_config_map_has_no_finalizers_or_timestamp_and_only_has_controller_owner_ref(rabbitmq)))
     );
 
     assert forall |s, s_prime| pre(s) && #[trigger] stronger_next(s, s_prime) implies pre(s_prime) || post(s_prime) by {
@@ -2744,260 +2773,6 @@ proof fn lemma_from_after_get_server_config_map_step_to_after_update_server_conf
     );
 }
 
-pub open spec fn sts_key_not_exists_or_delete_msg_in_flight_or_sts_has_current_ref(rabbitmq: RabbitmqClusterView) -> StatePred<RMQCluster> {
-    let sts_key = make_stateful_set_key(rabbitmq.object_ref());
-    |s: RMQCluster| {
-        ||| {
-            &&& s.resource_key_exists(sts_key)
-            &&& (
-                    (exists |msg: Message| {
-                        #[trigger] s.message_in_flight(msg)
-                        && msg.dst.is_KubernetesAPI()
-                        && msg.content.is_delete_request_with_key(sts_key)
-                    })
-                    || s.resource_obj_of(sts_key).metadata.owner_references == Some(seq![rabbitmq.controller_owner_ref()])
-            )
-        }
-        ||| !s.resource_key_exists(sts_key)
-    }
-}
-
-proof fn lemma_eventually_only_valid_stateful_set_exists(spec: TempPred<RMQCluster>, rabbitmq: RabbitmqClusterView)
-    requires
-        rabbitmq.well_formed(),
-        spec.entails(always(lift_state(RMQCluster::busy_disabled()))),
-        spec.entails(always(lift_action(RMQCluster::next()))),
-        spec.entails(tla_forall(|i| RMQCluster::kubernetes_api_next().weak_fairness(i))),
-        spec.entails(tla_forall(|i| RMQCluster::builtin_controllers_next().weak_fairness(i))),
-        spec.entails(always(lift_state(RMQCluster::desired_state_is(rabbitmq)))),
-        spec.entails(always(lift_state(helper_invariants::stateful_set_has_no_finalizers_or_timestamp_and_only_has_controller_owner_ref(rabbitmq)))),
-        spec.entails(always(lift_state(helper_invariants::every_update_sts_req_does_the_same(rabbitmq)))),
-        spec.entails(always(lift_state(helper_invariants::every_create_sts_req_does_the_same(rabbitmq)))),
-    ensures
-        spec.entails(true_pred().leads_to(always(lift_state(helper_invariants::stateful_set_has_owner_reference_pointing_to_current_cr(rabbitmq))))),
-{
-    let sts_key = make_stateful_set_key(rabbitmq.object_ref());
-    let key_not_exists = |s: RMQCluster| !s.resource_key_exists(sts_key);
-    let key_exists_and_current_ref = |s: RMQCluster| {
-        s.resource_key_exists(sts_key)
-        && s.resource_obj_of(make_stateful_set_key(rabbitmq.object_ref())).metadata.owner_references == Some(seq![rabbitmq.controller_owner_ref()])
-    };
-    let key_exists_and_old_ref = |s: RMQCluster| {
-        s.resource_key_exists(sts_key)
-        && exists |uid: nat| #![auto]
-        uid != rabbitmq.metadata.uid.get_Some_0() && s.resource_obj_of(make_stateful_set_key(rabbitmq.object_ref())).metadata.owner_references == Some(seq![OwnerReferenceView {
-            block_owner_deletion: None,
-            controller: Some(true),
-            kind: RabbitmqClusterView::kind(),
-            name: rabbitmq.metadata.name.get_Some_0(),
-            uid: uid,
-        }])
-    };
-    let delete_msg_in_flight = sts_key_not_exists_or_delete_msg_in_flight_or_sts_has_current_ref(rabbitmq);
-    implies_to_leads_to(spec, lift_state(key_not_exists), lift_state(delete_msg_in_flight));
-    implies_to_leads_to(spec, lift_state(key_exists_and_current_ref), lift_state(delete_msg_in_flight));
-    let post = helper_invariants::stateful_set_has_owner_reference_pointing_to_current_cr(rabbitmq);
-    let input = (BuiltinControllerChoice::GarbageCollector, sts_key);
-    let stronger_next = |s, s_prime: RMQCluster| {
-        &&& RMQCluster::next()(s, s_prime)
-        &&& RMQCluster::desired_state_is(rabbitmq)(s)
-        &&& helper_invariants::stateful_set_has_no_finalizers_or_timestamp_and_only_has_controller_owner_ref(rabbitmq)(s)
-        &&& helper_invariants::every_update_sts_req_does_the_same(rabbitmq)(s)
-        &&& helper_invariants::every_create_sts_req_does_the_same(rabbitmq)(s)
-    };
-    entails_always_and_n!(
-        spec,
-        lift_action(RMQCluster::next()),
-        lift_state(RMQCluster::desired_state_is(rabbitmq)),
-        lift_state(helper_invariants::stateful_set_has_no_finalizers_or_timestamp_and_only_has_controller_owner_ref(rabbitmq)),
-        lift_state(helper_invariants::every_update_sts_req_does_the_same(rabbitmq)),
-        lift_state(helper_invariants::every_create_sts_req_does_the_same(rabbitmq))
-    );
-    temp_pred_equality(
-        lift_action(stronger_next),
-        lift_action(RMQCluster::next())
-        .and(lift_state(RMQCluster::desired_state_is(rabbitmq)))
-        .and(lift_state(helper_invariants::stateful_set_has_no_finalizers_or_timestamp_and_only_has_controller_owner_ref(rabbitmq)))
-        .and(lift_state(helper_invariants::every_update_sts_req_does_the_same(rabbitmq)))
-        .and(lift_state(helper_invariants::every_create_sts_req_does_the_same(rabbitmq)))
-    );
-
-    assert forall |s, s_prime: RMQCluster| key_exists_and_old_ref(s) && #[trigger] stronger_next(s, s_prime) && RMQCluster::builtin_controllers_next().forward(input)(s, s_prime) implies delete_msg_in_flight(s_prime) by {
-        let owner_references = s.resource_obj_of(make_stateful_set_key(rabbitmq.object_ref())).metadata.owner_references.get_Some_0();
-        assert(owner_references.len() == 1);
-        let new_delete_msg = built_in_controller_req_msg(delete_req_msg_content(
-            sts_key, s.rest_id_allocator.allocate().1
-        ));
-        assert(s_prime.resource_key_exists(sts_key));
-        assert(s_prime.message_in_flight(new_delete_msg));
-    }
-
-    assert forall |s, s_prime: RMQCluster| key_exists_and_old_ref(s) && #[trigger] stronger_next(s, s_prime) implies key_exists_and_old_ref(s_prime) || delete_msg_in_flight(s_prime) by {
-        let step = choose |step| RMQCluster::next_step(s, s_prime, step);
-        match step {
-            Step::BuiltinControllersStep(i) => {
-                if i == input {
-                    assert(delete_msg_in_flight(s_prime));
-                } else {
-                    assert(key_exists_and_old_ref(s_prime));
-                }
-            },
-            Step::KubernetesAPIStep(i) => {
-                if i.get_Some_0().content.is_update_request_with_key(sts_key) {
-                    if RMQCluster::validate_update_request(i.get_Some_0().content.get_update_request(), s.kubernetes_api_state).is_Some()
-                    || RMQCluster::update_is_noop(i.get_Some_0().content.get_update_request().obj, s.resource_obj_of(sts_key)) {
-                        assert(key_exists_and_old_ref(s_prime));
-                    } else {
-                        assert(i.get_Some_0().content.get_update_request().obj.metadata.owner_references == Some(seq![rabbitmq.controller_owner_ref()]));
-                        StatefulSetView::spec_integrity_is_preserved_by_marshal();
-                        assert(!s_prime.resource_key_exists(sts_key) || (s_prime.resource_key_exists(sts_key) && s_prime.resource_obj_of(sts_key).metadata.owner_references == Some(seq![rabbitmq.controller_owner_ref()])));
-                    }
-                } else if i.get_Some_0().content.is_delete_request_with_key(sts_key) {
-                    assert(s.resource_obj_of(sts_key).metadata.finalizers.is_None());
-                    assert(!s_prime.resource_key_exists(sts_key));
-                } else {
-                    assert(key_exists_and_old_ref(s_prime));
-                }
-            },
-            _ => {
-                assert(key_exists_and_old_ref(s_prime) || delete_msg_in_flight(s_prime));
-            }
-        }
-    }
-
-    RMQCluster::lemma_pre_leads_to_post_by_builtin_controllers(
-        spec, input, stronger_next, RMQCluster::run_garbage_collector(), key_exists_and_old_ref, delete_msg_in_flight
-    );
-    or_leads_to_combine_n!(
-        spec,
-        lift_state(key_not_exists), lift_state(key_exists_and_current_ref), lift_state(key_exists_and_old_ref);
-        lift_state(delete_msg_in_flight)
-    );
-    implies_preserved_by_always_temp(
-        lift_state(helper_invariants::stateful_set_has_no_finalizers_or_timestamp_and_only_has_controller_owner_ref(rabbitmq)),
-        true_pred().implies(lift_state(key_not_exists).or(lift_state(key_exists_and_current_ref)).or(lift_state(key_exists_and_old_ref)))
-    );
-    entails_trans(
-        spec,
-        always(lift_state(helper_invariants::stateful_set_has_no_finalizers_or_timestamp_and_only_has_controller_owner_ref(rabbitmq))),
-        always(true_pred().implies(lift_state(key_not_exists).or(lift_state(key_exists_and_current_ref)).or(lift_state(key_exists_and_old_ref))))
-    );
-    implies_to_leads_to(spec, true_pred(), lift_state(key_not_exists).or(lift_state(key_exists_and_current_ref)).or(lift_state(key_exists_and_old_ref)));
-    lemma_delete_msg_in_flight_leads_to_only_valid_sts_exists(spec, rabbitmq);
-    leads_to_trans_n!(
-        spec,
-        true_pred(), lift_state(key_not_exists).or(lift_state(key_exists_and_current_ref)).or(lift_state(key_exists_and_old_ref)),
-        lift_state(delete_msg_in_flight),
-        lift_state(post)
-    );
-    leads_to_stable(spec, stronger_next, |s: RMQCluster| true, post);
-}
-
-proof fn lemma_delete_msg_in_flight_leads_to_only_valid_sts_exists(
-    spec: TempPred<RMQCluster>, rabbitmq: RabbitmqClusterView
-)
-    requires
-        rabbitmq.well_formed(),
-        spec.entails(always(lift_state(RMQCluster::busy_disabled()))),
-        spec.entails(always(lift_action(RMQCluster::next()))),
-        spec.entails(tla_forall(|i| RMQCluster::kubernetes_api_next().weak_fairness(i))),
-        spec.entails(tla_forall(|i| RMQCluster::builtin_controllers_next().weak_fairness(i))),
-        spec.entails(always(lift_state(RMQCluster::desired_state_is(rabbitmq)))),
-        spec.entails(always(lift_state(helper_invariants::stateful_set_has_no_finalizers_or_timestamp_and_only_has_controller_owner_ref(rabbitmq)))),
-        spec.entails(always(lift_state(helper_invariants::every_update_sts_req_does_the_same(rabbitmq)))),
-        spec.entails(always(lift_state(helper_invariants::every_create_sts_req_does_the_same(rabbitmq)))),
-    ensures
-        spec.entails(lift_state(sts_key_not_exists_or_delete_msg_in_flight_or_sts_has_current_ref(rabbitmq)).leads_to(lift_state(helper_invariants::stateful_set_has_owner_reference_pointing_to_current_cr(rabbitmq)))),
-{
-    let pre = sts_key_not_exists_or_delete_msg_in_flight_or_sts_has_current_ref(rabbitmq);
-    let post = helper_invariants::stateful_set_has_owner_reference_pointing_to_current_cr(rabbitmq);
-    let sts_key = make_stateful_set_key(rabbitmq.object_ref());
-    let key_exists_and_delete_msg_exists = |s: RMQCluster| {
-        &&& s.resource_key_exists(sts_key)
-        &&& exists |msg: Message| {
-            #[trigger] s.message_in_flight(msg)
-            && msg.dst.is_KubernetesAPI()
-            && msg.content.is_delete_request_with_key(sts_key)
-        }
-    };
-    let key_exists_and_current_ref = |s: RMQCluster| {
-        &&& s.resource_key_exists(sts_key)
-        &&& s.resource_obj_of(sts_key).metadata.owner_references == Some(seq![rabbitmq.controller_owner_ref()])
-    };
-    let key_not_exists = |s: RMQCluster| !s.resource_key_exists(sts_key);
-    valid_implies_implies_leads_to(spec, lift_state(key_exists_and_current_ref), lift_state(post));
-    valid_implies_implies_leads_to(spec, lift_state(key_not_exists), lift_state(post));
-
-    assert_by(
-        spec.entails(lift_state(key_exists_and_delete_msg_exists).leads_to(lift_state(post))),
-        {
-            let msg_to_p = |msg: Message| {
-                lift_state(|s: RMQCluster| {
-                    &&& s.resource_key_exists(sts_key)
-                    &&& s.message_in_flight(msg)
-                    &&& msg.dst.is_KubernetesAPI()
-                    &&& msg.content.is_delete_request_with_key(sts_key)
-                })
-            };
-            assert forall |msg: Message| spec.entails((#[trigger] msg_to_p(msg)).leads_to(lift_state(post))) by {
-                let input = Some(msg);
-                let msg_to_p_state = |s: RMQCluster| {
-                    &&& s.resource_key_exists(sts_key)
-                    &&& s.message_in_flight(msg)
-                    &&& msg.dst.is_KubernetesAPI()
-                    &&& msg.content.is_delete_request_with_key(sts_key)
-                };
-                let stronger_next = |s, s_prime: RMQCluster| {
-                    &&& RMQCluster::next()(s, s_prime)
-                    &&& RMQCluster::busy_disabled()(s)
-                    &&& helper_invariants::stateful_set_has_no_finalizers_or_timestamp_and_only_has_controller_owner_ref(rabbitmq)(s)
-                    &&& helper_invariants::every_update_sts_req_does_the_same(rabbitmq)(s)
-                };
-                entails_always_and_n!(
-                    spec,
-                    lift_action(RMQCluster::next()),
-                    lift_state(RMQCluster::busy_disabled()),
-                    lift_state(helper_invariants::stateful_set_has_no_finalizers_or_timestamp_and_only_has_controller_owner_ref(rabbitmq)),
-                    lift_state(helper_invariants::every_update_sts_req_does_the_same(rabbitmq))
-                );
-                temp_pred_equality(
-                    lift_action(stronger_next),
-                    lift_action(RMQCluster::next())
-                    .and(lift_state(RMQCluster::busy_disabled()))
-                    .and(lift_state(helper_invariants::stateful_set_has_no_finalizers_or_timestamp_and_only_has_controller_owner_ref(rabbitmq)))
-                    .and(lift_state(helper_invariants::every_update_sts_req_does_the_same(rabbitmq)))
-                );
-                RMQCluster::lemma_pre_leads_to_post_by_kubernetes_api(spec, input, stronger_next, RMQCluster::handle_request(), msg_to_p_state, post);
-            }
-            leads_to_exists_intro(spec, msg_to_p, lift_state(post));
-            assert_by(
-                tla_exists(msg_to_p) == lift_state(key_exists_and_delete_msg_exists),
-                {
-                    assert forall |ex| #[trigger] lift_state(key_exists_and_delete_msg_exists).satisfied_by(ex) implies tla_exists(msg_to_p).satisfied_by(ex) by {
-                        assert(ex.head().resource_key_exists(sts_key));
-                        let msg = choose |msg| {
-                            &&& #[trigger] ex.head().message_in_flight(msg)
-                            &&& msg.dst.is_KubernetesAPI()
-                            &&& msg.content.is_delete_request_with_key(sts_key)
-                        };
-                        assert(msg_to_p(msg).satisfied_by(ex));
-                    }
-                    temp_pred_equality(tla_exists(msg_to_p), lift_state(key_exists_and_delete_msg_exists));
-                }
-            )
-        }
-    );
-    temp_pred_equality(
-        lift_state(key_exists_and_delete_msg_exists).or(lift_state(key_exists_and_current_ref)).or(lift_state(key_not_exists)),
-        lift_state(pre)
-    );
-    or_leads_to_combine_n!(
-        spec,
-        lift_state(key_exists_and_delete_msg_exists), lift_state(key_exists_and_current_ref), lift_state(key_not_exists);
-        lift_state(post)
-    );
-}
-
 // Note that this variant only holds since the CR always exists and the old configmap owned by the old CR has been deleted.
 proof fn lemma_server_config_map_is_stable(
     spec: TempPred<RMQCluster>, rabbitmq: RabbitmqClusterView, p: TempPred<RMQCluster>
@@ -3007,7 +2782,7 @@ proof fn lemma_server_config_map_is_stable(
         spec.entails(always(lift_action(RMQCluster::next()))),
         spec.entails(always(lift_state(helper_invariants::no_delete_cm_req_is_in_flight(rabbitmq.object_ref())))),
         spec.entails(always(lift_state(helper_invariants::every_update_cm_req_does_the_same(rabbitmq)))),
-        spec.entails(always(lift_state(helper_invariants::server_config_map_has_no_finalizers_or_timestamp(rabbitmq)))),
+        spec.entails(always(lift_state(helper_invariants::server_config_map_has_no_finalizers_or_timestamp_and_only_has_controller_owner_ref(rabbitmq)))),
     ensures
         spec.entails(p.leads_to(always(lift_state(current_config_map_matches(rabbitmq))))),
 {
@@ -3016,21 +2791,21 @@ proof fn lemma_server_config_map_is_stable(
         &&& RMQCluster::next()(s, s_prime)
         &&& helper_invariants::no_delete_cm_req_is_in_flight(rabbitmq.object_ref())(s)
         &&& helper_invariants::every_update_cm_req_does_the_same(rabbitmq)(s)
-        &&& helper_invariants::server_config_map_has_no_finalizers_or_timestamp(rabbitmq)(s)
+        &&& helper_invariants::server_config_map_has_no_finalizers_or_timestamp_and_only_has_controller_owner_ref(rabbitmq)(s)
     };
     entails_always_and_n!(
         spec,
         lift_action(RMQCluster::next()),
         lift_state(helper_invariants::no_delete_cm_req_is_in_flight(rabbitmq.object_ref())),
         lift_state(helper_invariants::every_update_cm_req_does_the_same(rabbitmq)),
-        lift_state(helper_invariants::server_config_map_has_no_finalizers_or_timestamp(rabbitmq))
+        lift_state(helper_invariants::server_config_map_has_no_finalizers_or_timestamp_and_only_has_controller_owner_ref(rabbitmq))
     );
     temp_pred_equality(
         lift_action(stronger_next),
         lift_action(RMQCluster::next())
         .and(lift_state(helper_invariants::no_delete_cm_req_is_in_flight(rabbitmq.object_ref())))
         .and(lift_state(helper_invariants::every_update_cm_req_does_the_same(rabbitmq)))
-        .and(lift_state(helper_invariants::server_config_map_has_no_finalizers_or_timestamp(rabbitmq)))
+        .and(lift_state(helper_invariants::server_config_map_has_no_finalizers_or_timestamp_and_only_has_controller_owner_ref(rabbitmq)))
     );
 
     assert forall |s, s_prime| post(s) && #[trigger] stronger_next(s, s_prime) implies post(s_prime) by {
