+++ src/controller_examples/rabbitmq_controller/proof/liveness/mod.rs
@@ -2,3 +2,4 @@
 // SPDX-License-Identifier: MIT
 pub mod liveness;
 pub mod helper_invariants;
+pub mod terminate;
