+++ src/controller_examples/rabbitmq_controller/proof/liveness/terminate.rs
