+++ src/controller_examples/rabbitmq_controller/proof/mod.rs
@@ -3,4 +3,3 @@
 pub mod common;
 pub mod liveness;
 pub mod safety;
-pub mod terminate;
